AKI-24: Move modAion classes to modAionImpl
