+++ modAionImpl/test/org/aion/zero/impl/valid/EnergyLimitRuleTest.java
@@ -5,7 +5,7 @@ import static com.google.common.truth.Truth.assertThat;
 import java.util.ArrayList;
 import java.util.List;
 import org.aion.mcf.blockchain.valid.IValidRule;
-import org.aion.zero.api.BlockConstants;
+import org.aion.zero.impl.api.BlockConstants;
 import org.aion.zero.exceptions.HeaderStructureException;
 import org.aion.zero.types.A0BlockHeader;
 import org.junit.Test;
